+++ CHANGELOG.md
@@ -16,7 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.2.0]
 
 ### Added
-- Added new default footprint algorithm
+- Update default footprint algorithm, now there is only one algorithm for forge-py
+- Poetry update python libraries
 ### Deprecated
 ### Removed
 ### Fixed